chore: default to remote Supabase, drop web target, fix SDK drift, add DB CI
Point everyday development at the hosted Supabase project instead of
requiring the local stack (still optional for schema/RLS work and
pgTAP); local dev already ran against it. Drop unused Expo web support
(react-dom, react-native-web, web build config) since the app targets
iOS/Android only. Align Expo package versions with SDK 57 so
`expo-doctor` passes. Add a CI job that starts Supabase and runs
`supabase db reset` + `supabase test db`, plus `format:check`.

+++ supabase/tests/database/06_public_catalog.sql
@@ -41,7 +41,7 @@ $$, '22023', null, 'a non-positive guest count is rejected');
 select is(
   (select (result ->> 'totalAmountMinor')::bigint
      from public.search_available_units('ayni-hospitality', date '2027-03-01', date '2027-03-04', 2, 'es', '33333333-3333-3333-3333-333333333301') as result),
-  36000,
+  36000::bigint,
   'the price snapshot is nightly rate x nights in minor units (12000 x 3)');
 select is(
   (select (result ->> 'nights')::int
